wip: Try tejedor for local builds
This is the first step in experimenting with tejedor. This commit
adjusts the generate command to use a locally run instance of
tejedor to resolve dependencies. The package werkzeug is updated as
such.

The build-locally.sh script is then modified to use the tejedor proxy
and to allow binary content to be fetched by hermeto. (The script also
tweaks the requirements files so hermeto running in a container can
access the tejedor proxy running on the host system. This should be
temporary, maybe.)

The next step here is to adjust the Konflux build pipeline to spin up a
tejedor proxy so it can use it during build time.

Relates #893

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>