Build OCI artifact and streamline build pipeline
This commit drastically reduces the amount of tasks in the build
pipeline to clone + prefetch + build.

It also introduces a new task to build OCI artifacts instead of building
an image with a Containerfile via buildah. Besides being more efficient,
this will also help us, in the future, include the digest of the wheels
and sdist in the provenance.

This change also means the argfile.conf is no longer needed. They are
not removed as to avoid a mass rebuild event.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>